+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.0-beta6 (Unreleased)
 
-- (Enhancement) Add Bridge Group Interface support (`fmc_device_bridge_group_interface`)
+- (Enhancement) Add Bridge Group Interface (BVI) support (`fmc_device_bridge_group_interface`)
 
 ## 2.0.0-beta5